+++ lib/guilda_web/live/page_live.ex
@@ -5,13 +5,8 @@ defmodule GuildaWeb.PageLive do
   use GuildaWeb, {:live_view, container: {:div, class: "flex flex-col flex-grow"}}
 
   @impl true
-  def mount(params, session, socket) do
-    socket =
-      socket
-      |> assign_defaults(params, session)
-      |> assign(:page_title, page_title(socket.assigns.live_action))
-
-    {:ok, socket}
+  def mount(_params, _session, socket) do
+    {:ok, assign(socket, :page_title, page_title(socket.assigns.live_action))}
   end
 
   defp page_title(:index), do: "Home"
